Transfer database initialization to Repository
